(PC-23404)[PRO] refactor: remove Banner index and replace Banner import
