+++ pro/src/components/Banner/BannerInvisibleSiren.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Banner from 'ui-kit/Banners/Banner'
+import { Banner } from 'ui-kit'
 
 interface BannerInvisibleSirenProps {
   isNewOnboarding?: boolean
