+++ pro/src/components/Banner/BannerSummary.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { OFFER_WIZARD_MODE } from 'core/Offers/constants'
-import Banner from 'ui-kit/Banners/Banner'
+import { Banner } from 'ui-kit'
 
 interface BannerSummaryProps {
   mode: OFFER_WIZARD_MODE
