--- pro/src/ui-kit/Banners/Banner/index.ts
@@ -1,2 +1,0 @@
-export { default } from './Banner'
-export type { Link } from './Banner'
