+++ pro/src/ui-kit/index.ts
@@ -5,6 +5,6 @@ export { default as Tag } from './Tag'
 export * from './Button'
 export * from './form'
 export * from './typography'
-export { default as Banner } from './Banners/Banner'
+export { default as Banner } from './Banners/Banner/Banner'
 export { default as InfoBox } from './InfoBox'
 export { default as Thumb } from './Thumb'
